Unable to fetch new access token after expiry
The refreshTokens method expects a single parameter, however while calling this method from the aquireTokenSilent, two parameters are sent (refreshToken and scope) which leads to "Invalid Parameter: scope" error. 
This method invocation also has another issue, where "refreshToken" is sent as a parameter which a an object and not actually the refresh token which is required to fetch the new accessToken.